Added multiple components + tilemap movement works + readded UI + TPs
Animations implemented + edited map to fit new movement system + collision system for collectable objects + tag system + some code cleaning and refactoring

+++ Pacman/Vector2f.h
@@ -70,7 +70,7 @@ public:
 
 	float Vector2f::Length() const
 	{
-		return sqrt(X * X + Y * Y);
+		return sqrtf(X * X + Y * Y);
 	}
 
 	void Vector2f::Normalize()
